requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Cryptography
blake3==0.4.1

# Storage & Caching
redis==5.0.1

# HTTP Client for Image Fetching
httpx==0.26.0
requests==2.31.0

# Async Support
asyncio==3.4.3

# Azure (optional integrations)
azure-identity==1.17.1
azure-storage-blob==12.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-dotenv==1.0.0
